requirements.txt
python-pptx>=0.6.23
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=10.0.0
